Delay copying of default config until user wants to edit it
This removes all the puts/gets prompting about the config file on
initial startup (see Debian bug #523011). Instead, when the user first
presses F12, they are asked if they want a copy of the default config
to work on.

I haven't changed the actual messages, so they are currently too long.
Maybe a temporary buffer could pop up with the full text of the
question.